rename package plugins to extensions
