+++ src/main/java/org/gridsuite/mapping/server/common/extensions/SubtypesRegister.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.mapping.server.common.plugins;
+package org.gridsuite.mapping.server.common.extensions;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.Map;
