+++ src/main/java/org/gridsuite/mapping/server/dto/automata/extensions/AutomatonSubtypesExtension.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.mapping.server.dto.automata.plugins;
+package org.gridsuite.mapping.server.dto.automata.extensions;
 
 /**
  * @author Thang PHAM <quyet-thang.pham at rte-france.com>
