+++ src/main/java/org/gridsuite/mapping/server/dto/automata/extensions/AutomatonSubtypesRegister.java
@@ -4,9 +4,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.mapping.server.dto.automata.plugins;
+package org.gridsuite.mapping.server.dto.automata.extensions;
 
-import org.gridsuite.mapping.server.common.plugins.SubtypesRegister;
+import org.gridsuite.mapping.server.common.extensions.SubtypesRegister;
 import org.gridsuite.mapping.server.dto.automata.AbstractAutomaton;
 import org.gridsuite.mapping.server.model.AutomatonEntity;
